testshop/requirements.txt
asgiref==3.4.1
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-extra-views==0.13.0
django-haystack==3.1.1
django-oscar==3.1
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==9.6.0
idna==3.3
phonenumbers==8.12.35
Pillow==8.4.0
purl==1.6
pysolr==3.9.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7
